drukowanie listy

+++ administrator/components/com_repertoire/views/events/view.html.php
@@ -1,4 +1,5 @@
 <?php
+
 // No direct access to this file
 defined('_JEXEC') or die('Restricted access');
 
@@ -14,13 +15,23 @@ class RepertoireViewEvents extends JViewLegacy {
             return false;
         }
 
-        RepertoireHelper::addSubmenu('events');
-
+        $this->print = JRequest::getVar('print', false);
+        
         $this->addToolbar();
-        $this->sidebar = JHtmlSidebar::render();
+        if (!$this->print) {
+            RepertoireHelper::addSubmenu('events');
+            $this->sidebar = JHtmlSidebar::render();
+        }
 
-        // przypisanie zmiennych dla widoku z modelu
-        $this->rows = $this->get('Events');
+
+
+        $app = JFactory::getApplication();
+        if ($app->input->get('layout', 'default') == 'list') {
+            $this->rows = $this->getModel()->getSongs(JRequest::getVar('id'));
+        } else {
+            $this->rows = $this->get('Events');
+        }
+
 
         parent::display($tpl);
     }
@@ -29,10 +40,15 @@ class RepertoireViewEvents extends JViewLegacy {
         // tytuł strony
         JToolbarHelper::title(JText::_('COM_REPERTOIRE') . ': ' . JText::_('COM_REPERTOIRE_EVENTS'), 'stack article');
 
-        // przyciski
-        JToolBarHelper::addNew('event.add');
-        JToolBarHelper::editList('event.edit');
-        JToolBarHelper::deleteList(JText::_('COM_REPERTOIRE_CONFIRM_DELETE'), 'events.delete');
+        if (JFactory::getApplication()->input->get('layout', 'default') == 'list') {
+            JToolbarHelper::back();
+            JToolbarHelper::apply('event.toprint', 'COM_REPERTOIRE_PRINT');
+        } else {
+            // przyciski
+            JToolBarHelper::addNew('event.add');
+            JToolBarHelper::editList('event.edit');
+            JToolBarHelper::deleteList(JText::_('COM_REPERTOIRE_CONFIRM_DELETE'), 'events.delete');
+        }
         JToolbarHelper::preferences('com_repertoire');
     }
